Functional vehicle
Functional apart from rotate and closing ble connection.

+++ Vehicle_Simulator/app/src/main/java/team_project/matt/vehicle_simulator/VehicleStatus.java
@@ -2,9 +2,9 @@ package team_project.matt.vehicle_simulator;
 
 public interface VehicleStatus
 {
-    void notifyChargingStateChanged(boolean isCharging);
-    void notifyBatteryLevelChanged(double batteryLevel);
-    void notifyBatteryTemperatureChanged(double temperature);
-    void notifySpeedChanged(int speed);
-    void notifyDistanceChanged(int distance);
+    void      notifyChargingStateChanged(boolean isCharging);
+    void       notifyBatteryLevelChanged(int batteryLevel);
+    void notifyBatteryTemperatureChanged(int temperature);
+    void              notifySpeedChanged(int speed);
+    void           notifyDistanceChanged(int distance);
 }
